+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Underwater Game changelog
 
+## 1.17.0
+- Message all players when a player dies.
+- Prevent elements selection on webpage.
+- Reset all margin and padding for all webpage elements.
+- Add sound effect for a:
+  - hit missing,
+  - a hit landing,
+  - a very large hit landing,
+  - the player dying.
+
 ## 1.16.0
 - Give the player feedback on how much life she has.